% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listify.R
\name{listify}
\alias{listify}
\title{Turn vectors into lists with any specified linking word}
\usage{
listify(items, linking_word = "and", oxford_comma = FALSE)
}
\arguments{
\item{items}{A vector of items to turn into a list phrase (e.g. c("a", "b", "c")).}

\item{linking_word}{Defaults to "and". Can be anything.}

\item{oxford_comma}{\code{logical}. Defaults to \code{FALSE}. If TRUE, an oxford comma is added (e.g. "a, b, and c").}
}
\value{
A string in the form of a list (e.g. "a, b and c")
}
\description{
Turn vectors into lists with any specified linking word
}
\examples{
listify(c("a", "b", "c"), "or")

}
