% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_capitals.R
\name{restore_capitals}
\alias{restore_capitals}
\title{Restore sustom capitalisation in a string}
\usage{
restore_capitals(x, items_to_capitalise)
}
\arguments{
\item{x}{A string in which capitalisation needs to be restored}

\item{items_to_capitalise}{Whole words or acronyms in which capitalisation must be retained; special characters can be included (e.g. "R2-D2")}
}
\value{
A string with restored capitals
}
\description{
Restore sustom capitalisation in a string
}
\examples{
x <- "Should i tell c-3po the french call him z-6po?"
restore_capitals(x, c("I", "C-3PO", "French", "Z-6PO"))

}
